+++ hardwareStore/src/main/java/com/hardware/hardwareStore/model/Sale.java
@@ -20,7 +20,7 @@ public class Sale {
     @Column(name = "date_sale")
     private Date dateSale;
 
-    private Float total;
+    private Integer total;
     private String status;
 
     public Sale() {}
@@ -57,11 +57,11 @@ public class Sale {
         this.dateSale = dateSale;
     }
 
-    public Float getTotal() {
+    public Integer getTotal() {
         return total;
     }
 
-    public void setTotal(Float total) {
+    public void setTotal(Integer total) {
         this.total = total;
     }
